Run with `./setup.sh update`

+++ setup.sh
@@ -4,50 +4,51 @@
 cd "$( dirname "${BASH_SOURCE[0]}" )"
 
 #symlinks
+update=${1:-no}
 PWD=`pwd`
-PREFIX='.'
-DOTFILES=$(git ls-files)
-IGNOREFILES=( setup.sh README.md .gitignore )
+prefix='.'
+dotfiles=$(git ls-files)
+ignorefiles=( setup.sh README.md .gitignore )
 
 # Cleanup old installations
 
 # remove antigen
-if [ -e "${HOME}/.antigen" ]; then
+if [[ -e "${HOME}/.antigen" && "$update" == "no" ]]; then
     rm -rf ${HOME}/.antigen
 fi
 
 # remove zgen
-if [ -e "${HOME}/.zgen" ]; then
+if [[ -e "${HOME}/.zgen" && "$update" == "no" ]]; then
     rm -rf ${HOME}/.zgen
 fi
 
 # remove vim-plug
-if [ -e "${HOME}/.vim/autoload/plug.vim" ]; then
+if [[ -e "${HOME}/.vim/autoload/plug.vim" && "$update" == "no" ]]; then
     rm -f ${HOME}/.vim/autoload/plug.vim
 fi
 
-for DOTFILE in ${DOTFILES}; do
-    DOTFILEDIR=$(dirname ${DOTFILE})
+for dotfile in ${dotfiles}; do
+    dotfiledir=$(dirname ${dotfile})
 
-    for IGNOREFILE in ${IGNOREFILES[@]}; do
-        if [ ${DOTFILE} == ${IGNOREFILE} ]; then
+    for ignorefile in ${ignorefiles[@]}; do
+        if [[ "${dotfile}" == "${ignorefile}" ]]; then
             continue 2
         fi
     done
 
-    SYMLINK="${HOME}/${PREFIX}${DOTFILE}"
-    SYMLINKDIR="${HOME}/${PREFIX}${DOTFILEDIR}"
+    symlink="${HOME}/${prefix}${dotfile}"
+    symlinkdir="${HOME}/${prefix}${dotfiledir}"
 
-    if [ -e ${SYMLINK} ]; then
-        rm -f ${SYMLINK}
+    if [ -e ${symlink} ]; then
+        rm -f ${symlink}
     fi
 
-    echo "${PWD}/${DOTFILE} => ${SYMLINK}"
-    [[ -d ${SYMLINKDIR} ]] || mkdir -p ${SYMLINKDIR}
-    ln -sf ${PWD}/${DOTFILE} ${SYMLINK}
+    echo "${PWD}/${dotfile} => ${symlink}"
+    [[ -d ${symlinkdir} ]] || mkdir -p ${symlinkdir}
+    ln -sf ${PWD}/${dotfile} ${symlink}
 
-    if [ -e "${PWD}/${DOTFILE}.local" ]; then
-        echo "Adding local settings from ${PWD}/${DOTFILE}.local => ${SYMLINK}"
-        cat ${PWD}/${DOTFILE}.local >> ${SYMLINK}
+    if [[ -e "${PWD}/${dotfile}.local" && "$update" == no ]]; then
+        echo "Adding local settings from ${PWD}/${dotfile}.local => ${symlink}"
+        cat ${PWD}/${dotfile}.local >> ${symlink}
     fi
 done
